Change type of array "num_conditions_per_comparator" to uint32.

+++ python/boomer/common/rules.pxd
@@ -157,7 +157,7 @@ cdef class ModelBuilder:
     cdef void set_default_rule(self, Prediction* default_prediction)
 
     cdef void add_rule(self, Prediction* head, double_linked_list[Condition] conditions,
-                       intp[::1] num_conditions_per_comparator)
+                       uint32[::1] num_conditions_per_comparator)
 
     cdef RuleModel build_model(self)
 
@@ -179,6 +179,6 @@ cdef class RuleListBuilder(ModelBuilder):
     cdef void set_default_rule(self, Prediction* default_prediction)
 
     cdef void add_rule(self, Prediction* head, double_linked_list[Condition] conditions,
-                       intp[::1] num_conditions_per_comparator)
+                       uint32[::1] num_conditions_per_comparator)
 
     cdef RuleModel build_model(self)
